requirements.txt
fastapi
uvicorn
pydantic
Pillow
numpy
tensorflow
